# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Absen Ok API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/task/task-all": {
            "get": {
                "tags": [
                    "Approved Tasks"
                ],
                "summary": "Project Manager Access - Retrieve all to-do lists",
                "description": "Project Manager retrieves all task data",
                "responses": {
                    "200": {
                        "description": "Success: The task was successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Retrieve all to-do lists.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Retrieve all to-do lists.",
                                            "type": "string",
                                            "example": "The task was successfully retrieved."
                                        },
                                        "data": {
                                            "description": "Retrieve all to-do lists.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "description": "Retrieve all to-do lists.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Retrieve all to-do lists.",
                                                        "type": "string",
                                                        "example": "Tugas Pertama pada Project Pertama"
                                                    },
                                                    "project_id": {
                                                        "description": "Retrieve all to-do lists.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "deadline": {
                                                        "description": "Retrieve all to-do lists.",
                                                        "type": "string",
                                                        "example": "2023-11-27 10:50:10"
                                                    },
                                                    "description": {
                                                        "description": "Retrieve all to-do lists.",
                                                        "type": "text",
                                                        "example": "Ini adalah tugas pertama saya dari Project pertama"
                                                    },
                                                    "status": {
                                                        "description": "Retrieve all to-do lists.",
                                                        "type": "string",
                                                        "enum": [
                                                            "to-do",
                                                            "in progress",
                                                            "completed"
                                                        ],
                                                        "example": "completed"
                                                    },
                                                    "comment": {
                                                        "description": "Retrieve all to-do lists.",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "comment": {
                                                                    "description": "Retrieve all to-do lists.",
                                                                    "type": "text",
                                                                    "example": "Ini adalah komentar 1"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task/task-list": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Project Manager & Member Access - Get All Task",
                "description": "The user retrieves all task data",
                "operationId": "App\\Http\\Controllers\\Dashboard\\Task\\TaskController::getTaskList",
                "responses": {
                    "200": {
                        "description": "Success: All Task retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Get All Task.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Get All Task.",
                                            "type": "string",
                                            "example": "All Task retrieved successfully."
                                        },
                                        "data": {
                                            "description": "Get All Task.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "description": "Get All Task.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Get All Task.",
                                                        "type": "string",
                                                        "example": "Tugas Pertama pada Project Pertama"
                                                    },
                                                    "project_id": {
                                                        "description": "Get All Task.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "deadline": {
                                                        "description": "Get All Task.",
                                                        "type": "string",
                                                        "example": "2023-11-27 10:50:10"
                                                    },
                                                    "description": {
                                                        "description": "Get All Task.",
                                                        "type": "text",
                                                        "example": "Ini adalah tugas pertama saya dari Project pertama"
                                                    },
                                                    "status": {
                                                        "description": "Get All Task.",
                                                        "type": "string",
                                                        "enum": [
                                                            "to-do",
                                                            "in progress",
                                                            "completed"
                                                        ],
                                                        "example": "completed"
                                                    },
                                                    "comment": {
                                                        "description": "Get All Task.",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "comment": {
                                                                    "description": "Get All Task.",
                                                                    "type": "text",
                                                                    "example": "Ini adalah komentar 1"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Approved Tasks"
                ],
                "summary": "Project Manager Access - edit the status in the approved task table",
                "description": "The user retrieves all task data",
                "responses": {
                    "200": {
                        "description": "Success: status changed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "edit the status in the approved task table.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "edit the status in the approved task table.",
                                            "type": "string",
                                            "example": "status changed successfully."
                                        },
                                        "data": {
                                            "description": "edit the status in the approved task table.",
                                            "properties": {
                                                "user_id": {
                                                    "description": "edit the status in the approved task table.",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "task_id": {
                                                    "description": "edit the status in the approved task table.",
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "status": {
                                                    "description": "edit the status in the approved task table.",
                                                    "type": "string",
                                                    "example": "approved"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/task-pending": {
            "get": {
                "tags": [
                    "Approved Tasks"
                ],
                "summary": "Project Manager Access - Retrieve all tasks that have pending status",
                "description": "Project Manager retrieves all task data Pending",
                "responses": {
                    "200": {
                        "description": "Success: The task was successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Retrieve all tasks that have pending status.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Retrieve all tasks that have pending status.",
                                            "type": "string",
                                            "example": "The task was successfully retrieved."
                                        },
                                        "data": {
                                            "description": "Retrieve all tasks that have pending status.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "description": "Retrieve all tasks that have pending status.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Retrieve all tasks that have pending status.",
                                                        "type": "string",
                                                        "example": "Tugas Pertama pada Project Pertama"
                                                    },
                                                    "project_id": {
                                                        "description": "Retrieve all tasks that have pending status.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "deadline": {
                                                        "description": "Retrieve all tasks that have pending status.",
                                                        "type": "string",
                                                        "example": "2023-11-27 10:50:10"
                                                    },
                                                    "description": {
                                                        "description": "Retrieve all tasks that have pending status.",
                                                        "type": "text",
                                                        "example": "Ini adalah tugas pertama saya dari Project pertama"
                                                    },
                                                    "status": {
                                                        "description": "Retrieve all tasks that have pending status.",
                                                        "type": "string",
                                                        "enum": [
                                                            "to-do",
                                                            "in progress",
                                                            "completed"
                                                        ],
                                                        "example": "completed"
                                                    },
                                                    "comment": {
                                                        "description": "Retrieve all tasks that have pending status.",
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "comment": {
                                                                    "description": "Retrieve all tasks that have pending status.",
                                                                    "type": "text",
                                                                    "example": "Ini adalah komentar 1"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}